+++ README.md
@@ -1,1 +1,3 @@
 # fitness_app
+
+This is a simple fitness tracking React application. The code has been split into multiple files under the `src/` directory. The main entry point is `src/App.jsx` and components are organized into `src/components` and `src/screens`.
